chore: add pre-push hook for clippy/fmt enforcement
- Creates scripts/hooks/pre-push to prevent pushing code that fails CI
- Runs 'cargo fmt --check' and 'cargo clippy -- -D warnings'
- Copy to .git/hooks/pre-push to enable locally

+++ src/gateway/handlers.rs
@@ -685,7 +685,7 @@ pub async fn get_all_balances(
     })?;
 
     // Query all balances
-    match crate::funding::service::TransferService::get_all_balances(pg_db.pool(), user_id as i64)
+    match crate::funding::service::TransferService::get_all_balances(pg_db.pool(), user_id)
         .await
     {
         Ok(balances) => Ok((StatusCode::OK, Json(ApiResponse::success(balances)))),
